# Conflicts:
#	Jenkinsfile
#	README.md
#	dependencies.list
#	src/realm/object-store/impl/realm_coordinator.hpp
#	src/realm/object-store/impl/results_notifier.cpp
#	src/realm/object-store/impl/weak_realm_notifier.cpp
#	src/realm/object-store/impl/weak_realm_notifier.hpp
#	src/realm/object-store/object.cpp
#	src/realm/object-store/object.hpp
#	src/realm/object-store/object_accessor.hpp
#	src/realm/object-store/property.hpp
#	src/realm/object-store/sync/app.cpp
#	src/realm/object-store/sync/app.hpp
#	src/realm/object-store/sync/app_credentials.cpp
#	src/realm/object-store/sync/app_service_client.hpp
#	src/realm/object-store/sync/generic_network_transport.hpp
#	src/realm/object-store/sync/impl/sync_file.cpp
#	src/realm/object-store/sync/impl/sync_file.hpp
#	src/realm/object-store/sync/impl/sync_metadata.cpp
#	src/realm/object-store/sync/impl/sync_metadata.hpp
#	src/realm/object-store/sync/push_client.cpp
#	src/realm/object-store/sync/push_client.hpp
#	src/realm/object-store/sync/remote_mongo_collection.cpp
#	src/realm/object-store/sync/remote_mongo_collection.hpp
#	src/realm/object-store/sync/sync_manager.cpp
#	src/realm/object-store/sync/sync_manager.hpp
#	src/realm/object-store/sync/sync_session.cpp
#	src/realm/object-store/sync/sync_session.hpp
#	src/realm/object-store/sync/sync_user.cpp
#	src/realm/object-store/sync/sync_user.hpp
#	src/realm/object-store/thread_safe_reference.cpp
#	src/realm/object-store/thread_safe_reference.hpp
#	src/realm/object-store/util/apple/scheduler.hpp
#	src/realm/object-store/util/bson/bson.cpp
#	src/realm/object-store/util/generic/scheduler.hpp
#	src/realm/object-store/util/scheduler.cpp
#	src/realm/object-store/util/uv/scheduler.hpp
#	src/sync/impl/sync_client.hpp
#	src/sync/sync_config.hpp
#	test/object-store/migrations.cpp
#	test/object-store/object.cpp
#	test/object-store/primitive_list.cpp
#	test/object-store/realm.cpp
#	test/object-store/results.cpp
#	test/object-store/sync/file.cpp
#	test/object-store/sync/metadata.cpp
#	test/object-store/sync/remote_mongo_tests.cpp
#	test/object-store/sync/session/connection_change_notifications.cpp
#	test/object-store/sync/session/session.cpp
#	test/object-store/sync/session/wait_for_completion.cpp
#	test/object-store/sync/sync_manager.cpp
#	test/object-store/sync/user.cpp
#	test/object-store/thread_safe_reference.cpp
#	test/object-store/transaction_log_parsing.cpp
#	test/object-store/util/test_file.cpp
#	test/object-store/util/test_file.hpp
#	tests/bson.cpp
#	tests/sync/app.cpp
#	workflow/test_coverage.sh

+++ test/object-store/bson.cpp
@@ -374,7 +374,7 @@ TEST_CASE("canonical_extjson_corpus", "[bson]")
 
     SECTION("Double type")
     {
-        static double epsilon = 0.000000001;
+        static float epsilon = 0.000000001;
 
         SECTION("+1.0")
         {
